Final Fixes before merge

+++ com.unity.visualeffectgraph/VisualEffectGraph/Utilities/ParameterBinding/Library/VFXPlaneBinder.cs
@@ -37,13 +37,13 @@ namespace UnityEngine.Experimental.VFX.Utility
 
         public override bool IsValid(VisualEffect component)
         {
-            return Target != null && component.HasVector3((int)Position) && component.HasVector3((int)Normal);
+            return Target != null && component.HasVector3(Position) && component.HasVector3(Normal);
         }
 
         public override void UpdateBinding(VisualEffect component)
         {
-            component.SetVector3((int)Position, Target.transform.position);
-            component.SetVector3((int)Normal, Target.transform.up);
+            component.SetVector3(Position, Target.transform.position);
+            component.SetVector3(Normal, Target.transform.up);
         }
 
         public override string ToString()
